// jtrade/backend/src/core/user-project-platform/dto/user-project-platform-response.dto.ts
export class UserProjectPlatformResponseDto {
  id!: string;

  userId!: string;
  projectCodePlatformId!: string;

  isActive!: boolean;

  subscribedAt?: Date;
  lastDownloadAt?: Date | null;

  user?: {
    id: string;
    fullName?: string;
    email: string;
    role: string;
    avatarUrl?: string;
  };

  projectCodePlatform?: {
    id: string;
    deliveryMode?: string;
    runtimeMode?: string;
    status?: string;
    isActive: boolean;

    codeProject?: {
      id: string;
      projectKey: string;
      name: string;
      description?: string;
      isActive: boolean;

      typeProject?: {
        id: string;
        key: string;
        name: string;
        isActive: boolean;
      };
    };

    platform?: {
      id: string;
      name: string;
      category?: string;
      connectionType?: string;
      imageUrl?: string;
      isActive: boolean;
      isSupported: boolean;
    };

    companyProvider?: {
      id: string;
      companyName: string;
      legalName?: string;
      logoUrl?: string;
      isActive?: boolean;
    };
  };

  createdAt?: Date;
  updatedAt?: Date;
}